requirements.txt
django
django-storages
djangorestframework
pymysql
boto
requests
django-htmlmin
django-enumfield
